Move LazyRegex out of internal package
